% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootpval.R
\name{bootpval}
\alias{bootpval}
\title{Bootstrap P-Value Function}
\usage{
bootpval(x, conf.level = 0.95, iter = 3000, mu0 = 0, test = "two")
}
\arguments{
\item{x}{A numeric vector representing the sample data.}

\item{conf.level}{The confidence level for the bootstrap confidence interval (default is 0.95).}

\item{iter}{The number of bootstrap iterations (default is 3000).}

\item{mu0}{The null hypothesis mean (default is 0).}

\item{test}{The type of hypothesis test ("two" for two-tailed, "upper" for upper-tailed, "lower" for lower-tailed). Default is "two".}
}
\value{
A list containing the bootstrap p-value (\code{pvalue}), the test statistic (\code{tcalc}), the sample size (\code{n}), the original data (\code{x}), the test type (\code{test}), and the confidence interval (\code{ci}).
}
\description{
This function performs a bootstrap hypothesis test to calculate a p-value for the difference
between the sample mean and a specified null hypothesis mean (\code{mu0}). It generates bootstrap samples
and computes the test statistic for each sample, then calculates the p-value based on the resampled
statistics. It also provides a histogram plot of the bootstrap test statistics.
}
\examples{
set.seed(123)
sample_data <- rnorm(20, mean = 10, sd = 2)
result <- bootpval(x = sample_data, conf.level = 0.95, iter = 3000, mu0 = 10, test = "two")
print(result$pvalue)
}
