% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{fire Dataset}
\format{
A data frame with n rows and 2 columns:
\describe{
\item{DAMAGE}{Numeric: the amount of fire damage.}
\item{DISTANCE}{Numeric: the distance from the fire station.}
}
}
\source{
The source of your data (if applicable)
}
\usage{
data(fire)
}
\description{
A dataset containing information about fire damage.
}
\keyword{datasets}
