% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MATH4753NGUY0662.R
\name{mybin}
\alias{mybin}
\title{Binomial Simulation}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{The number of iterations (simulations) to perform. Default is 100.}

\item{n}{The number of trials in each iteration. Default is 10.}

\item{p}{The probability of success on each trial. Default is 0.5.}
}
\value{
A table showing the proportion of successes across all iterations.
}
\description{
Simulates binomial trials and generates a barplot of the relative frequencies of successes
over multiple iterations. This function takes a specified number of iterations and trials
and visualizes the distribution of successes in a barplot.
}
\examples{
# Run the binomial simulation with 100 iterations, 10 trials, and success probability of 0.7:
mybin(iter=100, n=10, p=0.7)

}
