% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MATH4753NGUY0662.R
\name{mycltp}
\alias{mycltp}
\title{Poisson Sampling Simulation}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n}{The number of samples per iteration (sample size). Default is 10.}

\item{iter}{The number of iterations (simulations). Default is 100.}

\item{lambda}{The rate parameter for the Poisson distribution. Default is 10.}

\item{...}{Additional graphical parameters to pass to the histogram plot.}
}
\value{
A plot showing the histogram of the sample means, a density curve, a theoretical normal curve, a barplot of the sampled data, and the Poisson probability mass function.
}
\description{
This function simulates random samples from a Poisson distribution, generates a histogram of the sample means, and adds a theoretical normal curve. It also displays a bar plot of the sampled data and the probability function of the Poisson distribution.
}
\examples{
# Example usage of mycltp function:
mycltp(n=20, iter=10000, lambda=10)

}
