% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{Maximum Likelihood Estimation with Plotting}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{A likelihood function that takes two arguments: an observed data vector \code{x} and a parameter value.}

\item{x}{A numeric vector of observed data.}

\item{param}{A numeric vector of possible parameter values to evaluate for maximum likelihood.}

\item{...}{Additional graphical parameters to pass to the plot.}
}
\value{
A list with elements:
\item{i}{The index of the maximum likelihood parameter value in \code{param}.}
\item{parami}{The parameter value that maximizes the likelihood.}
\item{yi}{The maximum likelihood value.}
\item{slope}{The slope around the maximum likelihood value (a vector), used to check for a maximum.}
}
\description{
This function computes the maximum likelihood estimate (MLE) of a parameter by evaluating
the likelihood function across a range of values and then identifying the value that maximizes the likelihood.
It also plots the likelihood function with an indication of the MLE.
}
